feat: step 2 rename folders
